+++ lib/chi-custom-elements/types/components.d.ts
@@ -5,16 +5,16 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "./stencil-public-runtime";
-import { AccordionSizes, IconSizes, LabelSizes, MarketingIconSizes, PickerPillSizes, PickerSizes, TabsSizes, TextInputSizes } from "./constants/size";
-import { AccordionItem, ChiMarketingIconModes, Country, DropdownMenuItem, DropdownSelectModes, FontWeight, FormWrapperCheckbox, FormWrapperRadio, GeneralPositions, PickerInputTypes, PickerOption, SearchInputModes, TabTrigger } from "./constants/types";
+import { AccordionSizes, IconSizes, LabelSizes, MarketingIconSizes, PickerPillSizes, PickerSizes, SkeletonSizes, TabsSizes, TextInputSizes } from "./constants/size";
+import { AccordionItem, ChiMarketingIconModes, Country, DropdownMenuItem, DropdownMenuMultiItem, DropdownSelectModes, FontWeight, FormWrapperCheckbox, FormWrapperRadio, GeneralPositions, PickerInputTypes, PickerOption, SearchInputModes, SkeletonTypes, TabTrigger } from "./constants/types";
 import { AlertColors, IconColors, TooltipColors } from "./constants/color";
 import { ChiStates } from "./constants/states";
 import { AppLayoutFormats, DataLocales, DateFormats, DatePickerModes, FormWrapperLayouts, FormWrapperTypes, TextInputTypes, TimePickerFormats, TimePickerTimeSteps } from "./constants/constants";
 import { Placement } from "popper.js";
 import { CountryCode } from "libphonenumber-js";
 import { ChiStates as ChiStates1 } from "./components";
-export { AccordionSizes, IconSizes, LabelSizes, MarketingIconSizes, PickerPillSizes, PickerSizes, TabsSizes, TextInputSizes } from "./constants/size";
-export { AccordionItem, ChiMarketingIconModes, Country, DropdownMenuItem, DropdownSelectModes, FontWeight, FormWrapperCheckbox, FormWrapperRadio, GeneralPositions, PickerInputTypes, PickerOption, SearchInputModes, TabTrigger } from "./constants/types";
+export { AccordionSizes, IconSizes, LabelSizes, MarketingIconSizes, PickerPillSizes, PickerSizes, SkeletonSizes, TabsSizes, TextInputSizes } from "./constants/size";
+export { AccordionItem, ChiMarketingIconModes, Country, DropdownMenuItem, DropdownMenuMultiItem, DropdownSelectModes, FontWeight, FormWrapperCheckbox, FormWrapperRadio, GeneralPositions, PickerInputTypes, PickerOption, SearchInputModes, SkeletonTypes, TabTrigger } from "./constants/types";
 export { AlertColors, IconColors, TooltipColors } from "./constants/color";
 export { ChiStates } from "./constants/states";
 export { AppLayoutFormats, DataLocales, DateFormats, DatePickerModes, FormWrapperLayouts, FormWrapperTypes, TextInputTypes, TimePickerFormats, TimePickerTimeSteps } from "./constants/constants";
@@ -315,6 +315,10 @@ export namespace Components {
          */
         "disabled": boolean;
         /**
+          * Displaying seconds column
+         */
+        "displaySeconds"?: boolean;
+        /**
           * To specify which dates to disable
          */
         "excludedDates": string;
@@ -478,6 +482,10 @@ export namespace Components {
           * To provide icon tooltip message
          */
         "iconTooltipMessage"?: string;
+        /**
+          * Dropdown menu items
+         */
+        "items"?: DropdownMenuMultiItem[];
         /**
           * To set position of the Dropdown
          */
@@ -1067,6 +1075,16 @@ export namespace Components {
           * To provide number of items in the dropdown to be displayed, and apply scroll if needed
          */
         "visibleItems"?: number;
+    }
+    interface ChiSkeleton {
+        /**
+          * Size of the skeleton: { xs, sm, md, lg, xl }.
+         */
+        "size": SkeletonSizes;
+        /**
+          * Applies type classes to the skeleton.
+         */
+        "type": SkeletonTypes;
     }
     interface ChiSpinner {
         /**
@@ -1906,6 +1924,12 @@ declare global {
         prototype: HTMLChiSearchInputElement;
         new (): HTMLChiSearchInputElement;
     };
+    interface HTMLChiSkeletonElement extends Components.ChiSkeleton, HTMLStencilElement {
+    }
+    var HTMLChiSkeletonElement: {
+        prototype: HTMLChiSkeletonElement;
+        new (): HTMLChiSkeletonElement;
+    };
     interface HTMLChiSpinnerElement extends Components.ChiSpinner, HTMLStencilElement {
     }
     var HTMLChiSpinnerElement: {
@@ -2056,6 +2080,7 @@ declare global {
         "chi-progress": HTMLChiProgressElement;
         "chi-radio-button": HTMLChiRadioButtonElement;
         "chi-search-input": HTMLChiSearchInputElement;
+        "chi-skeleton": HTMLChiSkeletonElement;
         "chi-spinner": HTMLChiSpinnerElement;
         "chi-switch": HTMLChiSwitchElement;
         "chi-tabs": HTMLChiTabsElement;
@@ -2384,6 +2409,10 @@ declare namespace LocalJSX {
          */
         "disabled"?: boolean;
         /**
+          * Displaying seconds column
+         */
+        "displaySeconds"?: boolean;
+        /**
           * To specify which dates to disable
          */
         "excludedDates"?: string;
@@ -2543,6 +2572,10 @@ declare namespace LocalJSX {
           * To provide icon tooltip message
          */
         "iconTooltipMessage"?: string;
+        /**
+          * Dropdown menu items
+         */
+        "items"?: DropdownMenuMultiItem[];
         /**
           * Triggered when hiding the Dropdown
          */
@@ -3241,6 +3274,16 @@ declare namespace LocalJSX {
          */
         "visibleItems"?: number;
     }
+    interface ChiSkeleton {
+        /**
+          * Size of the skeleton: { xs, sm, md, lg, xl }.
+         */
+        "size"?: SkeletonSizes;
+        /**
+          * Applies type classes to the skeleton.
+         */
+        "type"?: SkeletonTypes;
+    }
     interface ChiSpinner {
         /**
           * to render spinners on dark or light backgrounds.
@@ -3631,6 +3674,7 @@ declare namespace LocalJSX {
         "chi-progress": ChiProgress;
         "chi-radio-button": ChiRadioButton;
         "chi-search-input": ChiSearchInput;
+        "chi-skeleton": ChiSkeleton;
         "chi-spinner": ChiSpinner;
         "chi-switch": ChiSwitch;
         "chi-tabs": ChiTabs;
@@ -3673,6 +3717,7 @@ declare module "@stencil/core" {
             "chi-progress": LocalJSX.ChiProgress & JSXBase.HTMLAttributes<HTMLChiProgressElement>;
             "chi-radio-button": LocalJSX.ChiRadioButton & JSXBase.HTMLAttributes<HTMLChiRadioButtonElement>;
             "chi-search-input": LocalJSX.ChiSearchInput & JSXBase.HTMLAttributes<HTMLChiSearchInputElement>;
+            "chi-skeleton": LocalJSX.ChiSkeleton & JSXBase.HTMLAttributes<HTMLChiSkeletonElement>;
             "chi-spinner": LocalJSX.ChiSpinner & JSXBase.HTMLAttributes<HTMLChiSpinnerElement>;
             "chi-switch": LocalJSX.ChiSwitch & JSXBase.HTMLAttributes<HTMLChiSwitchElement>;
             "chi-tabs": LocalJSX.ChiTabs & JSXBase.HTMLAttributes<HTMLChiTabsElement>;