+++ src/documentation/pages/components/forms/text-input/_builder.vue
@@ -1,0 +1,7 @@
+<template lang="pug">
+  div
+    h2 Builder
+    ChiBuilder
+</template>
+
+<script lang="ts" setup></script>